This fixes VC++ compile error caused by unresolvable data types.  Doesn't cause an error (or warning even) in gcc.
git-svn-id: https://vtr-verilog-to-routing.googlecode.com/svn/trunk@1828 8e3573b8-cf2c-4f14-ef6d-137439e28b8b

+++ vpr/SRC/power/power_sizing.c
@@ -232,7 +232,7 @@ void power_sizing_init(t_arch * arch) {
 
 	// tech size = 2 lambda, so lambda^2/4.0 = tech^2
 	g_MTA_area = ((POWER_MTA_L * POWER_MTA_W)/ 4.0)*pow(g_power_tech->tech_size,
-			2.0);
+			(float)2.0);
 
 	// Determines physical size of different PBs
 	power_size_pb();
